Fix issue with Optimize Old 3DS keys

+++ source/utils.cpp
@@ -150,12 +150,12 @@ uint32_t read_be32(const uint8_t* data, size_t count, size_t multiplier) {
 
 uint64_t read_le64(const uint8_t* data, size_t count, size_t multiplier) {
 	data += count * multiplier;
-	return data[0] | (data[1] << 8) | (data[2] << 16) | (data[3] << 24) | (((uint64_t)data[4]) << 32) | (((uint64_t)data[5]) << 40) | (((uint64_t)data[6]) << 48) | (((uint64_t)data[7]) << 56);
+	return ((uint64_t)data[0]) | (((uint64_t)data[1]) << 8) | (((uint64_t)data[2]) << 16) | (((uint64_t)data[3]) << 24) | (((uint64_t)data[4]) << 32) | (((uint64_t)data[5]) << 40) | (((uint64_t)data[6]) << 48) | (((uint64_t)data[7]) << 56);
 }
 
 uint64_t read_be64(const uint8_t* data, size_t count, size_t multiplier) {
 	data += count * multiplier;
-	return data[7] | (data[6] << 8) | (data[5] << 16) | (data[4] << 24) | (((uint64_t)data[3]) << 32) | (((uint64_t)data[2]) << 40) | (((uint64_t)data[1]) << 48) | (((uint64_t)data[0]) << 56);
+	return ((uint64_t)data[7]) | (((uint64_t)data[6]) << 8) | (((uint64_t)data[5]) << 16) | (((uint64_t)data[4]) << 24) | (((uint64_t)data[3]) << 32) | (((uint64_t)data[2]) << 40) | (((uint64_t)data[1]) << 48) | (((uint64_t)data[0]) << 56);
 }
 
 void write_le16(uint8_t* data, uint16_t value, size_t count, size_t multiplier) {
